fix bug

+++ code/Camel.py
@@ -40,10 +40,10 @@ parser.add_argument('--batch_size', type = int, default = 64,
 parser.add_argument('--learn_rate', type = float, default = 0.001,
 				   help = 'learning rate')
 
-parser.add_argument('--train_iter_max', type = float, default = 1000,
+parser.add_argument('--train_iter_max', type = int, default = 1000,
 				   help = 'max number of training iterations')
 
-parser.add_argument('--save_model_freq', type = float, default = 5,
+parser.add_argument('--save_model_freq', type = int, default = 5,
 				   help = 'number of iterations to save model')
 
 parser.add_argument('--c_reg', type = float, default = 0.001,
